+++ src/ServiceDefinition.php
@@ -29,6 +29,11 @@ final class ServiceDefinition
     private $isFactory;
 
     /**
+     * @var bool
+     */
+    private $isAlias;
+
+    /**
      * @var array
      */
     private $arguments;
@@ -55,6 +60,7 @@ final class ServiceDefinition
         $this->class       = $this->className($name, $config);
         $this->isSingleton = isset($config['singleton']) ? $config['singleton'] : $singletonDefault;
         $this->isFactory   = isset($config['factory']);
+        $this->isAlias     = isset($config['service']);
         $this->arguments   = isset($config['arguments']) ? $config['arguments'] : [];
         $this->methods     = isset($config['methods']) ? $config['methods'] : [];
     }
@@ -92,6 +98,14 @@ final class ServiceDefinition
     }
 
     /**
+     * @return bool
+     */
+    public function isAlias()
+    {
+        return $this->isAlias;
+    }
+
+    /**
      * @return array
      */
     public function getArguments()
@@ -119,6 +133,18 @@ final class ServiceDefinition
     {
         if (isset($config['class']) && isset($config['factory'])) {
             throw InvalidConfigException::fromNameWhenClassAndFactorySpecified($name);
+        }
+
+        if (isset($config['class']) && isset($config['service'])) {
+            throw InvalidConfigException::fromNameWhenClassAndServiceSpecified($name);
+        }
+
+        if (isset($config['factory']) && isset($config['service'])) {
+            throw InvalidConfigException::fromNameWhenFactoryAndServiceSpecified($name);
+        }
+
+        if (isset($config['service'])) {
+            return $config['service'];
         }
 
         if (isset($config['class'])) {