+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased - TBD
+## 0.5.2 - 2016-09-30
 ### Added
   * Service aliases.
   * Container injection using `Configurator::container()`.